some small fixes for FreeBSD
git-svn-id: https://bugengine.googlecode.com/svn/trunk@667 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/engine/core/src/threads/platform=posix,darwin/thread.cc
@@ -113,7 +113,7 @@ u64 Thread::id() const
 
 u64 Thread::currentId()
 {
-    return (u64)pthread_self();
+    return u64(ptrdiff_t(pthread_self()));
 }
 
 istring Thread::name()
